feat: platform use slots hooks
