--- apps/api/v2/src/modules/slots/inputs/remove-selected-slot.input.ts
@@ -1,7 +1,0 @@
-import { IsOptional, IsString } from "class-validator";
-
-export class RemoveSelectedSlotInput {
-  @IsString()
-  @IsOptional()
-  uid?: string;
-}
