--- apps/api/v2/src/modules/slots/inputs/reserve-slot.input.ts
@@ -1,16 +1,0 @@
-import { IsDate, IsInt, IsOptional, IsString } from "class-validator";
-
-export class ReserveSlotInput {
-  @IsInt()
-  eventTypeId!: number;
-
-  @IsDate()
-  slotUtcStartDate!: string;
-
-  @IsDate()
-  slotUtcEndDate!: string;
-
-  @IsString()
-  @IsOptional()
-  bookingUid?: string;
-}
