+++ packages/platform/atoms/hooks/useReserveSlot.ts
@@ -1,0 +1,21 @@
+import { useQuery } from "@tanstack/react-query";
+
+import type { RemoveSelectedSlotInput, ApiResponse } from "@calcom/platform-types";
+
+import http from "../lib/http";
+
+export const QUERY_KEY = "delete-selected-slot";
+
+export const useReserveSlot = (props: RemoveSelectedSlotInput) => {
+  const deletedSlot = useQuery({
+    queryKey: [QUERY_KEY],
+    queryFn: () => {
+      return http
+        .post<ApiResponse<string>>("/slots/reserve", {
+          params: props,
+        })
+        .then((res) => res.data);
+    },
+  });
+  return deletedSlot;
+};
